requirements.txt
Flask>=3.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
joblib>=1.3.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
tldextract>=5.0.0
firebase-admin>=6.2.0
Flask-Login>=0.6.3
Flask-Session>=0.6.0
python-dotenv>=1.0.0
pyrebase4>=4.6.0
